add basic emulation of issue 2/3 $fe input behavior.
also save the last fe port state into the szx. (finally.)

+++ src/tape.h
@@ -81,5 +81,5 @@ TapePlayer_t *tape_player_from_tape(Tape_t *tape);
 void tape_player_close(TapePlayer_t *player);
 
 void tape_player_advance_cycles(TapePlayer_t *p, uint64_t cycles);
-uint8_t tape_player_get_next_sample(TapePlayer_t *player, uint64_t cycles);
+int tape_player_get_next_sample(TapePlayer_t *player, uint64_t cycles);
 void tape_player_pause(TapePlayer_t *player, bool paused);
